add categories to a bunch of stuff
