+++ examples/test_high_switch/TestHighSwitch.net
@@ -334,7 +334,7 @@
   (value "5600200520 (Molex)")
   (footprint "calisco:Molex_DuraClik_vert_5pin")
   (property (name "Sheetname") (value "can"))
-  (property (name "Sheetfile") (value "electronics_lib.CalSolBlocks.CalSolCanBlock"))
+  (property (name "Sheetfile") (value "examples.test_high_switch.CalSolCanBlock"))
   (property (name "edg_path") (value "can.conn"))
   (property (name "edg_short_path") (value "can.conn"))
   (property (name "edg_refdes") (value "J3"))
@@ -345,7 +345,7 @@
   (value "0ZCM0005FF2G")
   (footprint "Resistor_SMD:R_0603_1608Metric")
   (property (name "Sheetname") (value "can"))
-  (property (name "Sheetfile") (value "electronics_lib.CalSolBlocks.CalSolCanBlock"))
+  (property (name "Sheetfile") (value "examples.test_high_switch.CalSolCanBlock"))
   (property (name "edg_path") (value "can.can_fuse.fuse"))
   (property (name "edg_short_path") (value "can.can_fuse"))
   (property (name "edg_refdes") (value "F1"))
@@ -389,7 +389,7 @@
   (value "PESD1CAN,215")
   (footprint "Package_TO_SOT_SMD:SOT-23")
   (property (name "Sheetname") (value "can"))
-  (property (name "Sheetfile") (value "electronics_lib.CalSolBlocks.CalSolCanBlock"))
+  (property (name "Sheetfile") (value "examples.test_high_switch.CalSolCanBlock"))
   (property (name "edg_path") (value "can.esd"))
   (property (name "edg_short_path") (value "can.esd"))
   (property (name "edg_refdes") (value "U5"))
