+++ examples/test_multimeter.py
@@ -4,7 +4,7 @@ from typing import List, Dict
 from edg import *
 
 
-class ResistorMux(KiCadImportableBlock, GeneratorBlock):
+class ResistorMux(Interface, KiCadImportableBlock, GeneratorBlock):
   """Generates an array of resistors with one side muxed and the other end an array. Passive-typed.
   Specify an infinite resistance for an open circuit."""
   def symbol_pinning(self, symbol_name: str) -> Dict[str, BasePort]:
@@ -149,7 +149,7 @@ class MultimeterCurrentDriver(KiCadSchematicBlock, Block):
       })
 
 
-class FetPowerGate(KiCadSchematicBlock, Block):
+class FetPowerGate(PowerSwitch, KiCadSchematicBlock, Block):
   """A high-side PFET power gate that has a button to power on, can be latched
   on by an external signal, and provides the button output as a signal.
   """
