+++ examples/test_simon.py
@@ -3,7 +3,7 @@ import unittest
 from edg import *
 
 
-class DomeButtonConnector(FootprintBlock):
+class DomeButtonConnector(Connector, FootprintBlock):
   @init_in_parent
   def __init__(self) -> None:
     super().__init__()
